test: execution payload provider
