--- crates/e2e-test-utils/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "rollup-node-e2e-test-utils"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-exclude.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-# alloy
-alloy-primitives.workspace = true
-
-# reth
-reth-db-api = { git = "https://github.com/scroll-tech/reth.git", default-features = false }
-reth-db = { git = "https://github.com/scroll-tech/reth.git", default-features = false }
-reth-e2e-test-utils.workspace = true
-reth-network-peers.workspace = true
-reth-node-builder.workspace = true
-reth-node-core.workspace = true
-reth-provider.workspace = true
-reth-tasks.workspace = true
-
-# reth scroll
-reth-scroll-chainspec.workspace = true
-reth-scroll-node.workspace = true
-scroll-alloy-rpc-types-engine.workspace = true
-
-# misc
-eyre.workspace = true
