--- crates/e2e-test-utils/src/lib.rs
@@ -1,4 +1,0 @@
-//! End-to-end test helpers for the rollup node.
-
-/// Node related helpers.
-pub mod node;
