+++ crates/engine/Cargo.toml
@@ -17,14 +17,12 @@ alloy-primitives.workspace = true
 alloy-rpc-types-engine.workspace = true
 
 # scroll-alloy
-scroll-alloy-network.workspace = true
 scroll-alloy-provider.workspace = true
 scroll-alloy-rpc-types-engine.workspace = true
 
 # reth
 reth-network-api.workspace = true
-reth-payload-primitives = { git = "https://github.com/scroll-tech/reth.git" }
-reth-primitives = { workspace = true }
+reth-payload-primitives.workspace = true
 
 # reth-scroll
 reth-scroll-chainspec.workspace = true
