+++ crates/providers/Cargo.toml
@@ -38,11 +38,11 @@ tokio = { workspace = true, default-features = false }
 tracing.workspace = true
 
 [dev-dependencies]
+alloy-rpc-client.workspace = true
 eyre.workspace = true
+reth-payload-primitives.workspace = true
 reth-scroll-node.workspace = true
-reth-node-builder.workspace = true
-reth-node-core.workspace = true
-reth-rpc-server-types.workspace = true
+reth-e2e-test-utils.workspace = true
 scroll-db = { workspace = true, features = ["test-utils"] }
 scroll-network.workspace = true
 tokio = { workspace = true, features = ["macros"] }
